后端 修改逻辑bug  一天内关闭有开启 的设备  耗电量应该是累加的

+++ src/main/java/com/example/springboot3_2vue3/mapper/equipmapper/Deviceuse2Mapper.java
@@ -12,5 +12,5 @@ import java.util.List;
  **/
 public interface Deviceuse2Mapper extends BaseMapper<Deviceuse2> {
 
-    List<Deviceuse2> selectAlltodayById(int equipmentid,String date);
+    List<Deviceuse2> selectAlltodayById(Long equipmentid,String date);
 }
